requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.3.0=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.11.1=pypi_0
blas=1.0=mkl
bokeh=2.1.1=py36_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2022.07.19=h06a4308_0
cachetools=4.2.4=pypi_0
certifi=2021.5.30=py36h06a4308_0
charset-normalizer=2.0.12=pypi_0
click=8.0.4=pypi_0
cudatoolkit=11.3.1=h2bc3f7f_2
cupy-cuda111=9.6.0=pypi_0
cycler=0.11.0=pyhd3eb1b0_0
dataclasses=0.8=pyh4f3eec9_6
dbus=1.13.18=hb2f20db_0
decorator=5.1.1=pypi_0
docker-pycreds=0.4.0=pypi_0
einops=0.4.1=pypi_0
entrypoints=0.4=pypi_0
expat=2.4.9=h6a678d5_0
fastrlock=0.8=pypi_0
ffmpeg=4.3=hf484d3e_0
fontconfig=2.13.1=h6c09931_0
freetype=2.11.0=h70c0345_0
gitdb=4.0.9=pypi_0
gitpython=3.1.18=pypi_0
glib=2.69.1=h4ff587b_1
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
google=3.0.0=pypi_0
google-auth=2.6.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.44.0=pypi_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
icu=58.2=he6710b0_3
idna=3.4=pypi_0
importlib-metadata=4.8.3=pypi_0
intel-openmp=2022.1.0=h9e868ea_3769
ipykernel=5.5.6=pypi_0
ipython=7.16.3=pypi_0
ipython-genutils=0.2.0=pypi_0
jedi=0.17.2=pypi_0
jinja2=3.0.3=pyhd3eb1b0_0
jpeg=9e=h7f8727e_0
jupyter-client=7.1.2=pypi_0
jupyter-core=4.9.2=pypi_0
kiwisolver=1.3.1=py36h2531618_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libdeflate=1.8=h7f8727e_5
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libtiff=4.4.0=hecacb30_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h7f8727e_2
libuv=1.40.0=h7b6447c_0
libwebp-base=1.2.4=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxml2=2.9.14=h74e7548_0
lz4-c=1.9.3=h295c915_1
markdown=3.3.7=pypi_0
markupsafe=2.0.1=py36h27cfd23_0
matplotlib=3.2.2=0
matplotlib-base=3.2.2=py36hef1b27d_0
mkl=2020.2=256
mkl-service=2.3.0=py36he8ac12f_0
mkl_fft=1.3.0=py36h54f3939_0
mkl_random=1.1.1=py36h0573a6f_0
ncurses=6.3=h5eee18b_3
nest-asyncio=1.5.6=pypi_0
nettle=3.7.3=hbbd107a_1
ninja=1.11.1=pypi_0
numpy=1.19.2=py36h54aff64_0
numpy-base=1.19.2=py36hfa32c7d_0
oauthlib=3.2.0=pypi_0
olefile=0.46=py36_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=1.1.1q=h7f8727e_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.1.5=py36ha9443f7_0
parso=0.7.1=pypi_0
pathtools=0.1.2=pypi_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=8.3.1=py36h2c7a002_0
pip=21.2.2=py36h06a4308_0
promise=2.3=pypi_0
prompt-toolkit=3.0.32=pypi_0
protobuf=3.19.6=pypi_0
psutil=5.9.2=pypi_0
ptyprocess=0.7.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pygments=2.13.0=pypi_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py36h05f1152_2
python=3.6.13=h12debd9_1
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.10.2=py3.6_cuda11.3_cudnn8.2.0_0
pytorch-mutex=1.0=cuda
pytz=2021.3=pyhd3eb1b0_0
pyyaml=6.0=pypi_0
pyzmq=24.0.1=pypi_0
qt=5.9.7=h5867ecd_1
readline=8.1.2=h7f8727e_1
requests=2.27.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.8=pypi_0
sentry-sdk=1.9.10=pypi_0
setproctitle=1.2.3=pypi_0
setuptools=58.0.4=py36h06a4308_0
shortuuid=1.0.9=pypi_0
sip=4.19.8=py36hf484d3e_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.0=pypi_0
soupsieve=2.3.2.post1=pypi_0
sqlite=3.39.3=h5082296_0
tensorboard=2.8.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tk=8.6.12=h1ccaba5_0
torch-tb-profiler=0.3.1=pypi_0
torchaudio=0.10.2=py36_cu113
torchvision=0.11.3=py36_cu113
tornado=6.1=py36h27cfd23_0
traitlets=4.3.3=pypi_0
typing_extensions=4.1.1=pyh06a4308_0
urllib3=1.26.12=pypi_0
wandb=0.13.4=pypi_0
wcwidth=0.2.5=pypi_0
werkzeug=2.0.3=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.6=h5eee18b_0
yaml=0.2.5=h7b6447c_0
zipp=3.6.0=pypi_0
zlib=1.2.12=h5eee18b_3
zstd=1.5.2=ha4553b6_0
